requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
dnspython==2.7.0
frozenlist==1.6.0
idna==3.10
motor==3.7.1
multidict==6.4.4
numpy==2.2.6
pandas==2.2.3
propcache==0.3.1
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yarl==1.20.0